Split router into modules
