+++ web_app/routers/darbuotojai.py
@@ -1,0 +1,144 @@
+from fastapi import APIRouter, Request, Form, HTTPException, Depends
+from fastapi.responses import HTMLResponse, RedirectResponse, Response
+from fastapi.templating import Jinja2Templates
+
+import sqlite3
+from db import init_db
+from modules.audit import log_action, fetch_logs
+from modules.login import assign_role
+from modules.roles import Role
+from modules.constants import EU_COUNTRIES, EMPLOYEE_ROLES, DRIVER_NATIONALITIES
+from ..utils import ensure_columns, compute_limits, compute_busena, table_csv_response, get_db
+from ..auth import user_has_role, require_roles
+import datetime
+from datetime import date
+import pandas as pd
+
+router = APIRouter()
+templates = Jinja2Templates(directory="web_app/templates")
+
+# ---- Darbuotojai ----
+
+
+@router.get("/darbuotojai", response_class=HTMLResponse)
+def darbuotojai_list(request: Request):
+    return templates.TemplateResponse("darbuotojai_list.html", {"request": request})
+
+
+@router.get("/darbuotojai/add", response_class=HTMLResponse)
+def darbuotojai_add_form(
+    request: Request,
+    db: tuple[sqlite3.Connection, sqlite3.Cursor] = Depends(get_db),
+):
+    conn, cursor = db
+    grupes = [r[0] for r in cursor.execute("SELECT numeris FROM grupes").fetchall()]
+    data = {"imone": request.session.get("imone", "")}
+    return templates.TemplateResponse(
+        "darbuotojai_form.html",
+        {
+            "request": request,
+            "data": data,
+            "roles": EMPLOYEE_ROLES,
+            "grupes": grupes,
+        },
+    )
+
+
+@router.get("/darbuotojai/{did}/edit", response_class=HTMLResponse)
+def darbuotojai_edit_form(
+    did: int,
+    request: Request,
+    db: tuple[sqlite3.Connection, sqlite3.Cursor] = Depends(get_db),
+):
+    conn, cursor = db
+    row = cursor.execute("SELECT * FROM darbuotojai WHERE id=?", (did,)).fetchone()
+    if not row:
+        raise HTTPException(status_code=404, detail="Not found")
+    columns = [col[1] for col in cursor.execute("PRAGMA table_info(darbuotojai)")]
+    data = dict(zip(columns, row))
+    grupes = [r[0] for r in cursor.execute("SELECT numeris FROM grupes").fetchall()]
+    return templates.TemplateResponse(
+        "darbuotojai_form.html",
+        {
+            "request": request,
+            "data": data,
+            "roles": EMPLOYEE_ROLES,
+            "grupes": grupes,
+        },
+    )
+
+
+@router.post("/darbuotojai/save")
+def darbuotojai_save(
+    request: Request,
+    did: int = Form(0),
+    vardas: str = Form(...),
+    pavarde: str = Form(""),
+    pareigybe: str = Form(""),
+    el_pastas: str = Form(""),
+    telefonas: str = Form(""),
+    grupe: str = Form(""),
+    imone: str = Form(""),
+    aktyvus: str = Form(None),
+    db: tuple[sqlite3.Connection, sqlite3.Cursor] = Depends(get_db),
+):
+    conn, cursor = db
+    akt = 1 if aktyvus else 0
+    if did:
+        cursor.execute(
+            "UPDATE darbuotojai SET vardas=?, pavarde=?, pareigybe=?, el_pastas=?, telefonas=?, grupe=?, imone=?, aktyvus=? WHERE id=?",
+            (vardas, pavarde, pareigybe, el_pastas, telefonas, grupe, imone, akt, did),
+        )
+        action = "update"
+    else:
+        cursor.execute(
+            "INSERT INTO darbuotojai (vardas, pavarde, pareigybe, el_pastas, telefonas, grupe, imone, aktyvus) VALUES (?,?,?,?,?,?,?,?)",
+            (vardas, pavarde, pareigybe, el_pastas, telefonas, grupe, imone, akt),
+        )
+        did = cursor.lastrowid
+        action = "insert"
+    conn.commit()
+    log_action(conn, cursor, request.session.get("user_id"), action, "darbuotojai", did)
+    return RedirectResponse("/darbuotojai", status_code=303)
+
+
+@router.get("/api/darbuotojai")
+def darbuotojai_api(
+    request: Request,
+    db: tuple[sqlite3.Connection, sqlite3.Cursor] = Depends(get_db),
+):
+    conn, cursor = db
+    if user_has_role(request, cursor, Role.ADMIN):
+        cursor.execute("SELECT * FROM darbuotojai")
+    else:
+        cursor.execute(
+            "SELECT * FROM darbuotojai WHERE imone=?",
+            (request.session.get("imone"),),
+        )
+    rows = cursor.fetchall()
+    columns = [col[1] for col in cursor.execute("PRAGMA table_info(darbuotojai)")]
+    data = [dict(zip(columns, row)) for row in rows]
+    return {"data": data}
+
+
+@router.get("/api/darbuotojai.csv")
+def darbuotojai_csv(
+    request: Request,
+    db: tuple[sqlite3.Connection, sqlite3.Cursor] = Depends(get_db),
+):
+    conn, cursor = db
+    if user_has_role(request, cursor, Role.ADMIN):
+        cursor.execute("SELECT * FROM darbuotojai")
+    else:
+        cursor.execute(
+            "SELECT * FROM darbuotojai WHERE imone=?",
+            (request.session.get("imone"),),
+        )
+    rows = cursor.fetchall()
+    columns = [col[1] for col in cursor.execute("PRAGMA table_info(darbuotojai)")]
+    df = pd.DataFrame(rows, columns=columns)
+    csv_data = df.to_csv(index=False)
+    headers = {"Content-Disposition": "attachment; filename=darbuotojai.csv"}
+    return Response(content=csv_data, media_type="text/csv", headers=headers)
+
+
